moves header component to views
